refactor: add base view without the need of a field to work.
The idea is to be able to create page with etherpad without the
need to synchronize the content inside the CMS Database.
